fix(python): custom code conflict
Stainless-Generated-From: 50ac9cb74177f2395afcecde930d741975bdea4d

+++ src/cloudflare/types/radar/attacks/layer7/top/ase_origin_response.py
@@ -126,7 +126,7 @@ class Top0(BaseModel):
 
     origin_asn_name: str = FieldInfo(alias="originAsnName")
 
-    rank: float
+    rank: int
 
     value: str
